refactor(test_functions): migrate Oakley1D to YAML spec
Convert the Oakley1D test function from the legacy ABC-based system to a
YAML-driven spec as part of the Phase 2 fixed-dimension migration.

- Add `oakley_1d.yaml` spec file
- Strip ABC boilerplate from `oakley_1d.py`,
  keeping only `evaluate()`
- Fix matplotlib ylabel and xlabel to use raw strings
- Revise page title to descriptive-first form

Ref: #503
